Move BountyActions to new directory
