+++ packages/page-bounties/src/BountyActions/BountyClaimAction.tsx
@@ -10,8 +10,8 @@ import { useBounties } from '@polkadot/app-bounties/hooks';
 import { TxButton } from '@polkadot/react-components';
 import { useAccounts } from '@polkadot/react-hooks';
 
-import { isClaimable } from './helpers';
-import { useTranslation } from './translate';
+import { isClaimable } from '../helpers';
+import { useTranslation } from '../translate';
 
 interface Props {
   beneficiaryId: AccountId;
