+++ packages/page-bounties/src/BountyActions/index.tsx
@@ -6,12 +6,12 @@ import type { Balance, BlockNumber, BountyIndex, BountyStatus } from '@polkadot/
 
 import React, { useCallback, useMemo } from 'react';
 
+import { getBountyStatus } from '../helpers';
 import BountyClaimAction from './BountyClaimAction';
 import BountyCuratorProposedActions from './BountyCuratorProposedActions';
 import BountyInitiateVoting from './BountyInitiateVoting';
 import CloseBounty from './CloseBounty';
 import ExtendBountyExpiryAction from './ExtendBountyExpiryAction';
-import { getBountyStatus } from './helpers';
 import ProposeCuratorAction from './ProposeCuratorAction';
 
 interface Props {
